+++ Changelog.rst
@@ -6,6 +6,7 @@ Changelog
 5.1.0 (dev)
 -----------
  * Renamed ``Result`` and ``ResultMeta`` to ``Hit`` and ``HitMeta`` respectively
+ * ``Response`` now stores ``Search`` which it gets as first arg to ``__init__``
 
 5.0.0 (2016-11-04)
 ------------------